pl_predictor_backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
pandas==2.3.3
scikit-learn==1.7.2
numpy==2.3.3
python-dateutil==2.9.0
pytz==2025.2
scipy==1.16.2
joblib==1.5.2
threadpoolctl==3.6.0
asgiref==3.10.0
sqlparse==0.5.3
tzdata==2025.2
six==1.17.0
